+ add mapping get_flights of selected driver;
