+++ src/main/java/a/gleb/bus_station/controllers/AdminController.java
@@ -106,58 +106,4 @@ public class AdminController {
         typeBusRepo.save(bus);
         return "redirect:/administrations/administrator";
     }
-
-
-    @RequestMapping(value = "/add_driver", method = RequestMethod.GET)
-    public String adminAddDriverGet(Map<String, Object> model) {
-        return "addDriver";
-    }
-
-    @RequestMapping(value = "/add_driver", method = RequestMethod.POST)
-    public String adminAddDriverPost(@RequestParam String driverName,
-                                     @RequestParam String driverSurname,
-                                     @RequestParam String driverPhone,
-                                     Map<String, Object> model) {
-        Drivers driver = new Drivers(driverName, driverSurname, driverPhone);
-        driversRepo.save(driver);
-        return "redirect:/administrations/administrator/drivers";
-    }
-
-    @RequestMapping(value = "/administrator/drivers/{id}/edit", method = RequestMethod.GET)
-    public String administratorDriverEditGet(@PathVariable(value = "id") Integer id,
-                                             Map<String, Object> model) {
-        Optional<Drivers> driver = driversRepo.findById(id);
-        ArrayList<Drivers> driverModel = new ArrayList<>();
-        driver.ifPresent(driverModel::add);
-        model.put("driver", driverModel);
-        return "administratorEditDriver";
-    }
-
-    @RequestMapping(value = "/administrator/drivers/{id}/edit", method = RequestMethod.POST)
-    public String administratorDriverEditPost(@PathVariable(value = "id") Integer id,
-                                              @RequestParam String driverName,
-                                              @RequestParam String driverSurname,
-                                              @RequestParam String driverPhone,
-                                              Map<String, Object> model) {
-        int driverId = id;
-        Drivers driver = driversRepo.findById(driverId);
-        driver.setDriverName(driverName);
-        driver.setDriverSurname(driverSurname);
-        driver.setDriverPhone(driverPhone);
-        driversRepo.save(driver);
-        return "redirect:/administrations/administrator/drivers";
-    }
-
-    @RequestMapping(value = "/administrator/drivers/{id}/del", method = RequestMethod.GET)
-    public String administratorDriverDelete(@PathVariable(value = "id") Integer id,
-                                            Map<String, Object> model) {
-        int driverId = id;
-        Drivers driver = driversRepo.findById(driverId);
-        Iterable<BusFlights> flights = driver.getBusFlights();
-        for (BusFlights flight:flights) {
-            flight.setDrivers(driversRepo.findById(1));
-        }
-        driversRepo.delete(driver);
-        return "redirect:/administrations/administrator/drivers";
-    }
 }
