+++ src/main/resources/templates/administratorDriverFlights.html
@@ -1,0 +1,53 @@
+<!DOCTYPE html>
+<html lang="en">
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>Администратор: редактировать водителя</title>
+    <link rel="icon" href="/static/img/bus.png">
+</head>
+<body>
+<header th:insert="blocks/header.html :: header">
+</header>
+<main>
+    <div class="container">
+        <div class="row">
+            <div class="col s12">
+                <ul class="collection with-header">
+                    <li class="collection-header"><h4 th:each="driver : ${driver}" th:text="'Все рейсы водителя: ' + ${driver.driverSurname} + ' '"></h4></li>
+                </ul>
+                <table>
+                    <thead>
+                    <tr>
+                        <th>Тип маршрута:</th>
+                        <th>Пункт отправления:</th>
+                        <th>Пункт назначения:</th>
+                        <th>Дата:</th>
+                        <th>Подробнее:</th>
+                    </tr>
+                    </thead>
+
+                    <tbody>
+                    <tr th:each="el : ${flights}">
+                        <td th:text="${el.routeType}"/>
+                        <td th:text="${el.fromCity}">
+                        <td th:text="${el.toCity}"/>
+                        <td th:text="${el.dateFlight}"/>
+                        <td>
+                            <a th:href="'/flight/' + ${el.id}" class="waves-effect waves-light btn">Подробнее</a>
+                        </td>
+                    </tr>
+                    </tbody>
+                </table>
+            </div>
+        </div>
+
+    </div>
+
+
+</main>
+<script type="text/javascript" src="/static/js/materialize.min.js"></script>
+<footer th:insert="blocks/footer.html :: footer">
+</footer>
+</body>
+</html>
