+++ src/main/resources/templates/administratorDrivers.html
@@ -35,6 +35,7 @@
                     <th>Фамилия:</th>
                     <th>Имя:</th>
                     <th>Номер телефона:</th>
+                    <th>Рейсы:</th>
                     <th>Редактировать:</th>
                     <th>Удалить:</th>
                 </tr>
@@ -45,6 +46,9 @@
                     <td th:text="${el.driverSurname}"/>
                     <td th:text="${el.driverName}">
                     <td th:text="${el.driverPhone}"/>
+                    <td>
+                        <a th:href="'/administrations/administrator/driver/' + ${el.id} + '/get_flights'" class="waves-effect waves-light btn">Рейсы</a>
+                    </td>
                     <td>
                         <a th:href="'/administrations/administrator/drivers/' + ${el.id} + '/edit'" class="waves-effect waves-light btn">Редактировать</a>
                     </td>
